% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_doc.R
\name{disp_doc}
\alias{disp_doc}
\title{Display/show documentation associated with a dataset}
\usage{
disp_doc(x)
}
\arguments{
\item{x}{path/directory to dataset (e.g., md/csv/txt file)}
}
\value{
prints data documentation to the console
}
\description{
Display/show a markdown documentation associated with a dataset (e.g., csv, txt file). Documentation should have been generated by the doc_data() function.
}
\examples{
\dontrun{disp_doc("mcars.csv")}
}
\author{
Hause Lin
}
