% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\name{doc_data}
\alias{doc_data}
\title{doc_data function}
\usage{
doc_data(x, overwrite = FALSE)
}
\arguments{
\item{x}{path/directory to dataset}

\item{overwrite}{whether to overwrite documentation (defaults to FALSE)}
}
\description{
Generates documentation for a dataset
}
\examples{
\donttest{
doc_data("mtcars.csv")
}
}
\author{
Hause Lin
}
