% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_open.R
\name{doc_open}
\alias{doc_open}
\title{Open documentation markdown file in R or RStudio}
\usage{
doc_open(md_file)
}
\arguments{
\item{md_file}{markdown document filepath}
}
\description{
Opens a markdown documentation file in R or RStudio. A wrapper function for utils::file.exists()
}
\examples{
\dontrun{doc_open("mcars.md")}
}
\author{
Hause Lin
}
