% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_refresh.R
\name{doc_refresh}
\alias{doc_refresh}
\title{Refreshes and updates existing documentation for a dataset}
\usage{
doc_refresh(x)
}
\arguments{
\item{x}{filepath (e.g., md/csv/txt file)}
}
\description{
Update/refresh existing documentation that is outdated (maybe because dataset has been updated) or if you need to tidy up your existing documentation.
}
\examples{
\dontrun{doc_refresh("mcars.csv")}
}
\author{
Hause Lin
}
