% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\name{doc_data}
\alias{doc_data}
\title{Generate documentation in markdown for a dataset}
\usage{
doc_data(x, overwrite = FALSE)
}
\arguments{
\item{x}{dataset filepath (e.g., csv/txt file)}

\item{overwrite}{whether to overwrite documentation (defaults to FALSE)}
}
\value{
markdown file with the same name as input x in the same directory
}
\description{
Generates documentation for a dataset
}
\examples{
\dontrun{doc_data("./data/mcars.csv")}
}
\author{
Hause Lin
}
